完成核心框架优化的第一步(通过DB存储固化关键数据)

+++ argus-lib/src/com/lynx/lib/core/dex/PluginManager.java
@@ -1,23 +1,24 @@
 package com.lynx.lib.core.dex;
 
-import android.content.Context;
+import java.io.File;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+import android.text.TextUtils;
+
 import com.lynx.lib.core.Const;
 import com.lynx.lib.core.LFApplication;
 import com.lynx.lib.core.Logger;
-import com.lynx.lib.core.dex.DexModuleLoader.DexStatus;
-import com.lynx.lib.core.dex.DexModuleLoader.DexType;
+import com.lynx.lib.core.dex.DexLoader.DexStatus;
+import com.lynx.lib.db.DBService;
 import com.lynx.lib.http.HttpCallback;
 import com.lynx.lib.http.HttpService;
 import com.lynx.lib.http.core.HttpParam;
 import com.lynx.lib.util.FileUtil;
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * 
@@ -27,27 +28,21 @@ import java.util.Map;
  */
 public class PluginManager {
 	private static final String Tag = "pluginManager";
-	public static final String PREFIX = "plugincenter";
-	public static final String MY_PLUGIN_CONFIG = "myplugin_config";
-
 	private static final String LM_API_PLUGIN_CONFIG = "/dex/myplugin";
 
-	private Context context;
+	private LFApplication context;
 	private HttpService httpService;
+	private DBService dbService;
+
 	private Map<String, PluginLoader> pluginLoaders = new HashMap<String, PluginLoader>();
-	private JSONArray jaMyPlugins;
-	private String basicDir;
 
 	public PluginManager() {
 		this.context = LFApplication.instance();
 		this.httpService = (HttpService) LFApplication.instance().service("http");
+		this.dbService = context.dbService();
 
-		File tmp = new File(context.getFilesDir(), PREFIX);
-		if (!tmp.exists()) {
-			tmp.mkdirs();
-		}
-		basicDir = tmp.getAbsolutePath();
 		loadLocalPlugins();
+		deleteNoUseFolder();
 		updateCheck();
 	}
 
@@ -62,14 +57,14 @@ public class PluginManager {
 					return;
 				}
 				JSONArray jaPlugin = joResult.getJSONArray("data");
-				saveConfig(MY_PLUGIN_CONFIG, jaPlugin);
-
 				// 获取插件更新配置
 				for (int i = 0; i < jaPlugin.length(); ++i) {
 					try {
-						Plugin plugin = (Plugin) DexUtil.json2dexModule(DexType.PLUGIN,
-								jaPlugin.getJSONObject(i));
-						update(plugin);
+						Plugin plugin = context.gson().fromJson(
+								jaPlugin.getJSONObject(i).toString(), Plugin.class);
+						PluginLoader loader = pluginLoaders.get(plugin.getModule());
+						// TODO: 提示用户有插件更新
+						loader.update(plugin);
 					} catch (Exception e) {
 						Logger.e(Tag, "获取动态模块更新配置内容错误", e);
 					}
@@ -86,12 +81,12 @@ public class PluginManager {
 		}
 	};
 
-	private DexModuleListener installListener = new DexModuleListener() {
+	private DexListener installListener = new DexListener() {
 		@Override
-		public void onStatusChanged(DexModule dexModule, int status) {
+		public void onStatusChanged(DexLoader dexLoader, int status) {
 			switch (status) {
-			case DexModuleListener.DEX_INSTALL_FAIL:
-				removePluginLoader((Plugin) dexModule, null);
+			case DexListener.DEX_INSTALL_FAIL:
+				removePluginLoader((Plugin) dexLoader.dexModule(), null);
 			}
 		}
 	};
@@ -100,13 +95,13 @@ public class PluginManager {
 	 * 插件更新检查
 	 */
 	public void updateCheck() {
-		if (localPlugins() == null) {
+		if (TextUtils.isEmpty(localPlugins())) {
 			return;
 		}
 
 		HttpParam param = new HttpParam();
 		param.put("ua", "android");
-		param.put("apps", localPlugins());
+		param.put("plugins", localPlugins());
 		httpService.post(String.format("%s%s", Const.LM_API_DOMAIN, LM_API_PLUGIN_CONFIG), param,
 				callback);
 	}
@@ -122,58 +117,45 @@ public class PluginManager {
 	/**
 	 * 加入UI插件
 	 */
-	public void addPluginLoader(PluginLoader loader, DexModuleListener listener) {
+	public void addPluginLoader(Plugin plugin, DexListener listener) {
 		try {
+			PluginLoader loader = new PluginLoader(plugin, DexStatus.UPDATE);
+
 			loader.addListener(listener);
 			loader.addListener(installListener);
-			if (jaMyPlugins == null) {
-				jaMyPlugins = new JSONArray();
-			}
-			JSONObject jo = DexUtil.dexModule2json(DexType.PLUGIN, loader.dexModule());
-			jaMyPlugins.put(jo);
-			saveConfig(MY_PLUGIN_CONFIG, jaMyPlugins);
+
+			dbService.save(loader.dexModule());
 			pluginLoaders.put(loader.module(), loader);
 		} catch (Exception e) {
 			Logger.e(Tag, "add pluginloader error", e);
-			removePluginLoader((Plugin) loader.dexModule(), null);
+			removePluginLoader(plugin, null);
 		}
 	}
 
-	public void removePluginLoader(Plugin plugin, DexModuleListener listener) {
+	public void removePluginLoader(Plugin plugin, DexListener listener) {
+		PluginLoader loader = pluginLoaders.remove(plugin.getModule());
+
+		if (loader == null) {
+			Logger.e(Tag, String.format("no such pluginloader[%s] find", plugin.getModule()));
+			return;
+		}
+
 		try {
-			JSONArray ja = new JSONArray();
-			PluginLoader loader = pluginLoaders.get(plugin.module());
 			loader.addListener(listener);
 			loader.delete();
-			for (String key : pluginLoaders.keySet()) {
-				if (key.equals(plugin.module())) {
-					continue;
-				}
-				loader = pluginLoaders.get(key);
-				JSONObject jo = DexUtil.dexModule2json(DexType.PLUGIN, loader.dexModule());
-				ja.put(jo);
-			}
 
-			jaMyPlugins = ja;
-			saveConfig(MY_PLUGIN_CONFIG, jaMyPlugins);
-			pluginLoaders.remove(plugin.module());
+			String where = String.format("module='%s'", loader.module());
+			dbService.deleteByWhere(Plugin.class, where);
+			pluginLoaders.remove(loader.module());
+
 			if (listener != null) {
-				listener.onStatusChanged(plugin, DexModuleListener.DEX_UNINSTALL_SUCCESS);
+				listener.onStatusChanged(loader, DexListener.DEX_UNINSTALL_SUCCESS);
 			}
 		} catch (Exception e) {
 			Logger.e(Tag, "remove pluginloader error", e);
 			if (listener != null) {
-				listener.onStatusChanged(plugin, DexModuleListener.DEX_UNINSTALL_FAIL);
+				listener.onStatusChanged(loader, DexListener.DEX_UNINSTALL_FAIL);
 			}
-		}
-	}
-
-	// TODO: 提示用户有插件更新
-	private void update(Plugin plugin) {
-		PluginLoader dexLoader = getPluginLoader(plugin.module());
-		if (dexLoader != null) {
-			// UI module有新的动态更新
-			dexLoader.hasUpdate(plugin);
 		}
 	}
 
@@ -181,114 +163,39 @@ public class PluginManager {
 	 * 载入本地插件
 	 */
 	private void loadLocalPlugins() {
-		File path = new File(basicDir + File.separator + MY_PLUGIN_CONFIG);
-		if (path.length() == 0)
-			return;
-		FileInputStream fis = null;
 		try {
-			fis = new FileInputStream(path);
-			byte[] bytes = new byte[fis.available()];
-			fis.read(bytes);
-			fis.close();
-			String str = new String(bytes, "UTF-8");
-			JSONArray jaMyPlugins = new JSONArray(str);
-			if (jaMyPlugins == null || jaMyPlugins.length() <= 0) {
-				return;
-			}
-
-			for (int i = 0; i < jaMyPlugins.length(); ++i) {
-				Plugin plugin = (Plugin) DexUtil.json2dexModule(DexType.PLUGIN,
-						jaMyPlugins.getJSONObject(i));
-				if (plugin != null) {
-					PluginLoader loader = new PluginLoader(plugin, DexStatus.UPDATE);
-					pluginLoaders.put(loader.module(), loader);
-				}
+			List<Plugin> myPlugins = dbService.findAll(Plugin.class);
+			for (Plugin plugin : myPlugins) {
+				PluginLoader loader = new PluginLoader(plugin, DexStatus.LOAD);
+				pluginLoaders.put(loader.module(), loader);
 			}
 		} catch (Exception e) {
 			Logger.e(Tag, "load local plugins error", e);
-		} finally {
-			if (fis != null) {
-				try {
-					fis.close();
-				} catch (Exception e) {
-
-				}
-			}
 		}
 	}
 
 	private String localPlugins() {
 		String plugins = "";
 		for (String key : pluginLoaders.keySet()) {
-			plugins = String.format("%s|%s", plugins, key);
+			plugins = String.format("%s|%s,%s", plugins, key, pluginLoaders.get(key).version());
 		}
-
-		deleteNoUseFolder();
 
 		return plugins.length() > 1 ? plugins.substring(1) : null;
-	}
-
-	/**
-	 * 将配置文件保存到本地
-	 * 
-	 * @param data
-	 * @throws Exception
-	 */
-	private void saveConfig(String fileName, Object data) throws Exception {
-		File config = new File(basicDir, fileName);
-		File configTmp = new File(basicDir, "tmp");
-		File configOld = new File(basicDir, "old");
-		FileOutputStream fos = null;
-
-		if (configOld.exists()) {
-			configOld.delete();
-		}
-		if (config.exists()) {
-			config.renameTo(configOld);
-		}
-
-		try {
-			byte[] bytes = data.toString().getBytes("UTF-8");
-			fos = new FileOutputStream(configTmp);
-			fos.write(bytes);
-			fos.close();
-			fos = null;
-			config.delete();
-			if (!configTmp.renameTo(config)) {
-				// revert to old config file
-				if (config.exists()) {
-					config.delete();
-				}
-				configOld.renameTo(config);
-				throw new Exception("unable to move config from " + configTmp + " to " + config);
-			}
-		} catch (Exception e) {
-			config.delete();
-			configOld.renameTo(config);
-			Logger.e(Tag, "save config error", e);
-		} finally {
-			if (fos != null) {
-				try {
-					fos.close();
-				} catch (Exception e) {
-				}
-			}
-			configTmp.delete();
-			configOld.delete();
-		}
 	}
 
 	/**
 	 * 删除卸载插件的残余文件
 	 */
 	private void deleteNoUseFolder() {
-		String pluginBaseDir = "/data/data/app.name/files/plugin";
+		String pluginBaseDir = String.format("%s/plugin", context.getFilesDir().getAbsolutePath());
 		File pluginBase = new File(pluginBaseDir);
 		if (pluginBase.exists()) {
 			File[] files = pluginBase.listFiles();
 			for (File file : files) {
 				try {
-					FileUtil.deleteFile(file);
+					if (!pluginLoaders.keySet().contains(file.getName())) {
+						FileUtil.deleteFile(file);
+					}
 				} catch (Exception e) {
 					Logger.w(Tag, "删除卸载插件的残余文件异常", e);
 				}
